Use the fs.FS interface in the local library
This opens the path for mocking the file system in tests later down the
line.

+++ src/library/local_library.go
@@ -128,6 +128,7 @@ type LocalLibrary struct {
 
 	artFinder art.Finder
 
+	fs         fs.FS
 	sqlFilesFS fs.FS
 
 	imageScaler *scaler.Scaler
@@ -151,8 +152,7 @@ func (lib *LocalLibrary) Close() {
 // AddLibraryPath adds a library directory to the list of libraries which will be
 // scanned and consequently watched.
 func (lib *LocalLibrary) AddLibraryPath(path string) {
-	_, err := os.Stat(path)
-
+	_, err := lib.statFile(path)
 	if err != nil {
 		log.Printf("error adding path: %s", err)
 		return
@@ -403,9 +403,7 @@ func (lib *LocalLibrary) AddMedia(filename string) error {
 		return nil
 	}
 
-	_, err := os.Stat(filename)
-
-	if err != nil {
+	if _, err := lib.statFile(filename); err != nil {
 		return err
 	}
 
@@ -929,7 +927,7 @@ func (lib *LocalLibrary) Initialize() error {
 
 	// This database is already created and populated. We could just apply the
 	// migrations without executing the initial schema.
-	if st, err := os.Stat(lib.database); err == nil && st.Size() > 0 {
+	if st, err := lib.statFile(lib.database); err == nil && st.Size() > 0 {
 		return lib.applyMigrations()
 	}
 
@@ -995,6 +993,12 @@ func (lib *LocalLibrary) Truncate() error {
 		return nil
 	}
 
+	// The local library is not working with the actual file system. This is probably
+	// a mock file system for tests. So skip removing the database.
+	if _, ok := lib.fs.(*osFS); !ok {
+		return nil
+	}
+
 	return os.Remove(lib.database)
 }
 
@@ -1028,6 +1032,23 @@ func (lib *LocalLibrary) scaleImage(
 	return newBytesReadCloser(res), nil
 }
 
+func (lib *LocalLibrary) statFile(name string) (fs.FileInfo, error) {
+	if _, ok := lib.fs.(*osFS); ok {
+		// The library is using the os package directly for working with
+		// the file system so one syscall could be saved by directly sending
+		// stat instead of going through the fs.FS interface which requires
+		// open and stat to get to the fs.FileInfo.
+		return os.Stat(name)
+	}
+
+	fh, err := lib.fs.Open(name)
+	if err != nil {
+		return nil, err
+	}
+	defer fh.Close()
+	return fh.Stat()
+}
+
 // NewLocalLibrary returns a new LocalLibrary which will use for database the file
 // specified by databasePath. Also creates the database connection so you does not
 // need to worry about that. It accepts the parent's context and create its own
@@ -1040,6 +1061,7 @@ func NewLocalLibrary(
 	lib := new(LocalLibrary)
 	lib.database = databasePath
 	lib.sqlFilesFS = sqlFilesFS
+	lib.fs = &osFS{}
 
 	libContext, cancelFunc := context.WithCancel(ctx)
